Modifications to the Login and signup components of the page.
Now there is a (non)functional login form on the page, which at the moment does nothing
Additionally, there is a (semi)functional register form on the page, which simply sends and api request to the server passing in all of the information on the field.

Lastly, this push comes with some changes as to how Users are created. The user class has a static variable which increments whenever the User constructor is called.
This is to ensure that no two users are created that have an identical UID.
The administrator should be the first registered user in the system- which to be done on startup of the system and the database.

There is still much to be done, but this is the ground work for what we will need for Friday's deadline expectations of a working prototype.

Cameron Dickie
Functionality and adhering to the UML Design