System Structure Changes
This is the first step in moving the static list of users and courses from the ServerState class over to the SystemData class so that the SystemData observer can notify the admin of any changes.

Additionally, allowed for the display of the usertype on the front end scripted in from the server.
All that is left to load in on the front end is the course information directly from the server.

Cameron Dickie
Server Structure and minor front-end changes