Admin Application for Registration
Implemented the admin to have the option to accept or decline user registration requests

This makes use of the socket to update the client list of applications using the observer pattern

When a user goes to register now, instead of creating a user it creates an application, which when the admin accepts this application it creates the user with that information

Only once the application has been accepted can the user log in with their credentials.

Currently, there is no way for the user to know that their application has been accepted, but something to implement if we have time.

Cameron Dickie
Observer Pattern, Registration Overhaul, Front-end Changes