Added Course Creation on Front end
Added a route to create a course with a name, coursecode, professor and max students.

Currently does not support prerequisites for a course or student registration

This calls updateAll() for the systemdata subject, causing the admin to request a new set of courses and it automatically appears in their list of courses.

Cameron Dickie
Front and Back end functionality, course creation